split up response_parser
